[feat/#59] list 페이지 SSR 방식 변경
