+++ src/app/list/page.tsx
@@ -1,168 +1,13 @@
-"use client";
-
-import Image from "next/image";
-import styles from "./page.module.scss";
-import smallLogo from "@/assets/small_logo.svg";
-import ArrowButton from "@/components/common/Button";
-import { useRouter } from "next/navigation";
-import { Dropdown } from "@/components/common/Dropdown";
-import { UserCard } from "@/components/common/UserCard";
-import { Pagination } from "@/components/common/Pagination";
-import { useEffect, useRef, useState } from "react";
 import { getSubjects } from "@/services/subjects/getSubjects";
-import { Subjects } from "@/types/Subjects";
-import ListModal from "@/components/Modal/ListModal";
+import ListClientPage from "@/components/ListClientPage";
 
-const ListPage = () => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [sortedList, setSortedList] = useState<Subjects[]>([]);
-  const [personalList, setPersonalList] = useState<Subjects[]>([]);
-  const [sortedOption, setSortedOption] = useState("최신순");
-  const [storedId, setStoredId] = useState<Number[]>([]);
-  const [showModal, setShowModal] = useState(false);
-  const router = useRouter();
-  const [limitSize, setLimitSize] = useState(8);
-  const wrapperRef = useRef<HTMLUListElement | null>(null);
+const ListPage = async () => {
+  const limit = 8;
+  const offset = 0;
 
-  useEffect(() => {
-    const handleResize = () => {
-      if (!wrapperRef.current) return;
+  const { results } = await getSubjects(limit, offset);
 
-      const wrapperWidth = wrapperRef.current.offsetWidth;
-      const gap = 20;
-      const cardMinWidth = 186;
-      const cardCount = Math.floor((wrapperWidth + gap) / (cardMinWidth + gap));
-
-      const adjustedCardCount = Math.max(3, Math.min(cardCount, 4));
-      setLimitSize(adjustedCardCount * 2);
-    };
-
-    handleResize();
-    window.addEventListener("resize", handleResize);
-    return () => window.removeEventListener("resize", handleResize);
-  }, []);
-
-  useEffect(() => {
-    const stored = JSON.parse(localStorage.getItem("personalId") || "[]");
-    setStoredId(stored);
-  }, []);
-
-  useEffect(() => {
-    const personal = sortedList.filter((item) => {
-      return storedId.includes(item.id);
-    });
-    setPersonalList(personal);
-  }, [sortedList, storedId]);
-
-  const fetchAllSubjects = async () => {
-    let offset = 0;
-    let allData: Subjects[] = [];
-
-    while (true) {
-      const { results } = await getSubjects(limitSize, offset);
-      allData = [...allData, ...results];
-
-      if (results.length < limitSize) {
-        break;
-      }
-
-      offset += limitSize;
-    }
-
-    const sorted = sortedSubjects(allData, sortedOption);
-    setSortedList(sorted);
-    setCurrentPage(1);
-  };
-
-  useEffect(() => {
-    fetchAllSubjects();
-  }, [sortedOption]);
-
-  const sortedSubjects = (data: Subjects[], option: string) => {
-    const copy = [...data];
-    if (option === "이름순") {
-      return copy.sort((a, b) => a.name.localeCompare(b.name));
-    } else {
-      return copy.sort(
-        (a, b) =>
-          new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
-      );
-    }
-  };
-
-  const paginatedList = sortedList.slice(
-    (currentPage - 1) * limitSize,
-    currentPage * limitSize
-  );
-
-  const handlePageChange = (id: number) => {
-    if (storedId.includes(id)) {
-      router.push(`/post/${id}/answer`);
-    } else {
-      router.push(`/post/${id}`);
-    }
-  };
-
-  const handleOpenModal = () => {
-    setShowModal(true);
-  };
-
-  return (
-    <main className={styles["list-page"]}>
-      {showModal && (
-        <ListModal personalList={personalList} onModalChange={setShowModal} />
-      )}
-      <header className={styles["top-wrapper"]}>
-        <button
-          className={styles["logo-button"]}
-          onClick={() => router.push("/")}
-        >
-          <Image
-            src={smallLogo}
-            alt="로고"
-            width={146}
-            height={57}
-            className={styles.logo}
-          />
-        </button>
-        <ArrowButton mode="answer" showArrow={true} onClick={handleOpenModal}>
-          답변하러 가기
-        </ArrowButton>
-      </header>
-      <div className={styles['dropdown-wrapper']}>
-      <h1 className={styles.title}>누구에게 질문할까요?</h1>
-      <div className={styles.dropdown}>
-        <Dropdown
-          sortedOption={sortedOption}
-          setSortedOption={setSortedOption}
-        />
-      </div>
-      </div>
-      <ul ref={wrapperRef} className={styles["user-card-wrapper"]}>
-        {paginatedList.map(({ id, imageSource, name, questionCount }) => (
-          <li key={id}>
-            <button
-              className={styles["user-card-button"]}
-              onClick={() => handlePageChange(id)}
-            >
-              <UserCard
-                img={imageSource}
-                nickname={name}
-                question={questionCount}
-              />
-            </button>
-          </li>
-        ))}
-      </ul>
-      <nav className={styles["pagination"]}>
-        <Pagination
-          pageSize={limitSize}
-          currentPage={currentPage}
-          onPageChage={setCurrentPage}
-        />
-      </nav>
-    </main>
-  );
+  return <ListClientPage initialSubjects={results} limit={limit} />;
 };
 
 export default ListPage;
