+++ src/components/ListClientPage/index.module.scss
@@ -1,0 +1,101 @@
+@use "@/styles/_color.scss" as *;
+@use "@/styles/_typography.scss" as *;
+@use "@/styles/_shadow.scss" as *;
+@use "@/styles/_media.scss" as *;
+
+.list-page {
+  padding-top: 45px;
+  padding-left: 130px;
+  padding-right: 130px;
+
+  @include tablet {
+    padding-left: 50px;
+    padding-right: 50px;
+  }
+
+  @include mobile {
+    padding-left: 24px;
+    padding-right: 24px;
+  }
+}
+
+.top-wrapper {
+  display: flex;
+  justify-content: space-between;
+
+  @include mobile {
+    flex-direction: column;
+    align-items: center;
+    row-gap: 20px;
+  }
+}
+
+.logo-button,
+.user-card-button {
+  border: none;
+  background-color: transparent;
+  cursor: pointer;
+  margin: 0;
+  padding: 0;
+}
+
+.logo {
+  display: block;
+}
+
+.dropdown-wrapper {
+  display: flex;
+  flex-direction: column;
+  row-gap: 20px;
+  margin-top: 35px;
+  margin-bottom: 30px;
+
+  @include mobile {
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+  }
+}
+
+.title {
+  @include heading-1;
+  display: flex;
+  justify-content: center;
+  margin: 0;
+
+  @include mobile {
+    @include heading-3;
+    margin-top: 0;
+  }
+}
+
+.dropdown {
+  display: flex;
+  justify-content: center;
+}
+
+.user-card-wrapper {
+  display: grid;
+  // grid-template-columns: repeat(4, 1fr);
+  grid-template-columns: repeat(auto-fit, minmax(186px, 1fr));
+  grid-template-rows: repeat(2, 1fr);
+  gap: 20px;
+  padding: 0;
+  max-width: 940px;
+  margin: 0 auto;
+  list-style: none;
+
+  @include tablet {
+    width: 100%;
+  }
+}
+
+.user-card-button {
+  width: 100%;
+}
+
+.pagination {
+  display: flex;
+  justify-content: center;
+  margin-top: 30px;
+}
